Split test_permission_group_queries.py test file
